datadog-sandbox/flask-app/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
ddtrace==1.19.0
datadog==0.47.0
requests==2.31.0
